!24 support double in C Module * Merge branch 'master' into dev * fix float->double for bytecodegen * update to /package * Use double in C for float in python

+++ package/PikaMath/PikaMath_Operator.c
@@ -9,32 +9,32 @@ int PikaMath_Operator_NOT(PikaObj* self, int flag) {
 int PikaMath_Operator_OR(PikaObj* self, int flag1, int flag2) {
     return flag1 || flag2;
 }
-int PikaMath_Operator_equalFloat(PikaObj* self, float num1, float num2) {
+int PikaMath_Operator_equalFloat(PikaObj* self, double num1, double num2) {
     return num1 == num2;
 }
 int PikaMath_Operator_equalInt(PikaObj* self, int num1, int num2) {
     return num1 == num2;
 }
-int PikaMath_Operator_graterThanFloat(PikaObj* self, float num1, float num2) {
+int PikaMath_Operator_graterThanFloat(PikaObj* self, double num1, double num2) {
     return num1 > num2;
 }
 int PikaMath_Operator_graterThanInt(PikaObj* self, int num1, int num2) {
     return num1 > num2;
 }
-int PikaMath_Operator_lessThanFloat(PikaObj* self, float num1, float num2) {
+int PikaMath_Operator_lessThanFloat(PikaObj* self, double num1, double num2) {
     return num1 < num2;
 }
 int PikaMath_Operator_lessThanInt(PikaObj* self, int num1, int num2) {
     return num1 < num2;
 }
-float PikaMath_Operator_minusFloat(PikaObj* self, float num1, float num2) {
+double PikaMath_Operator_minusFloat(PikaObj* self, double num1, double num2) {
     return num1 - num2;
 }
 int PikaMath_Operator_minusInt(PikaObj* self, int num1, int num2) {
     return num1 - num2;
 }
 
-float PikaMath_Operator_plusFloat(PikaObj* self, float num1, float num2) {
+double PikaMath_Operator_plusFloat(PikaObj* self, double num1, double num2) {
     return num1 + num2;
 }
