+++ package/PikaStdLib/PikaStdLib_MemChecker.c
@@ -14,10 +14,10 @@ void PikaStdLib_MemChecker_resetMax(PikaObj* self) {
     pikaMemMaxReset();
 }
 
-float PikaStdLib_MemChecker_getMax(PikaObj* self) {
+double PikaStdLib_MemChecker_getMax(PikaObj* self) {
     return pikaMemMax() / 1024.0;
 }
 
-float PikaStdLib_MemChecker_getNow(PikaObj* self) {
+double PikaStdLib_MemChecker_getNow(PikaObj* self) {
     return pikaMemNow() / 1024.0;
 }
