+++ package/pika_cjson/cJSON.c
@@ -2573,7 +2573,7 @@ CJSON_PUBLIC(cJSON *) cJSON_CreateIntArray(const int *numbers, int count)
     return a;
 }
 
-CJSON_PUBLIC(cJSON *) cJSON_CreateFloatArray(const float *numbers, int count)
+CJSON_PUBLIC(cJSON *) cJSON_CreateFloatArray(const double *numbers, int count)
 {
     size_t i = 0;
     cJSON *n = NULL;
